feat(gateway): gateway preparation
Delete the previous structure. Prepare the gateway module with middleware. Prepare the
authentication module with the grpc client and connect to the user service. Prepare the approved
microservices management module and secure its path. Supplementary measures.

BREAKING CHANGE: The previous structure is completely erased and the new changes are included along
with the original file changes.